requirements.txt
google-generativeai==0.7.2
streamlit
python-dotenv==1.0.1
openai == 0.27.0
